.history/requirements_20210807213703.txt
dateparser
#PyPDF4~=1.27.0
fastapi
starlette
python-multipart
Pillow
uvicorn
pytesseract
pdf2image